+++ test/Psr7ResponseTest.php
@@ -60,6 +60,30 @@ class Psr7ResponseTest extends TestCase
     /**
      * @dataProvider getResponseData
      */
+    public function testResponseToZendWithMemoryStream($body, $status, $headers)
+    {
+        $stream = new Stream('php://memory', 'wb+');
+        $stream->write($body);
+
+        $psr7Response = new Response($stream, $status, $headers);
+        $this->assertInstanceOf(ResponseInterface::class, $psr7Response);
+
+        $zendResponse = Psr7Response::toZend($psr7Response);
+        $this->assertInstanceOf(ZendResponse::class, $zendResponse);
+        $this->assertEquals($body, (string)$zendResponse->getBody());
+        $this->assertEquals($status, $zendResponse->getStatusCode());
+
+        $zendHeaders = $zendResponse->getHeaders()->toArray();
+        foreach ($headers as $type => $values) {
+            foreach ($values as $value) {
+                $this->assertContains($value, $zendHeaders[$type]);
+            }
+        }
+    }
+
+    /**
+     * @dataProvider getResponseData
+     */
     public function testResponseToZendFromRealStream($body, $status, $headers)
     {
         $stream = new Stream(tempnam(sys_get_temp_dir(), 'Test'), 'wb+');